requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aioredis==2.0.1
python-dotenv==1.0.0
requests==2.31.0
asyncpg==0.28.0
tortoise==0.1.1
tortoise-orm==0.19.3
pandas==2.0.3